+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add nflreadr to dependecies and drop lubridate and magrittr dependency
 * The function `load_pbp()` now calls `nflreadr::load_pbp()` internally. Therefore the argument `qs` has been deprecated. It will be removed in a future release.
 * The deprecated arguments `source` and `pp` in the functions `fast_scraper_*()` and `build_nflfastR_pbp()` have been removed
+* Added the variables `racr` ("Receiver Air Conversion Ratio"), `target_share`, `air_yards_share`, `wopr` ("Weighted Opportunity Rating") and `pacr` ("Passing Air Conversion Ratio") to the output of `calculate_player_stats()`
 
 # nflfastR 4.2.0